Introduce TimeTableRepository to make it transactional
