+++ optaplanner-integration/optaplanner-spring-boot-example/src/main/java/org/optaplanner/spring/boot/example/TimeTableSpringBootApp.java
@@ -22,9 +22,9 @@ import java.time.LocalTime;
 import org.optaplanner.spring.boot.example.domain.Lesson;
 import org.optaplanner.spring.boot.example.domain.Room;
 import org.optaplanner.spring.boot.example.domain.Timeslot;
-import org.optaplanner.spring.boot.example.service.LessonRepository;
-import org.optaplanner.spring.boot.example.service.RoomRepository;
-import org.optaplanner.spring.boot.example.service.TimeslotRepository;
+import org.optaplanner.spring.boot.example.persistence.LessonRepository;
+import org.optaplanner.spring.boot.example.persistence.RoomRepository;
+import org.optaplanner.spring.boot.example.persistence.TimeslotRepository;
 import org.springframework.boot.CommandLineRunner;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
