+++ optaplanner-integration/optaplanner-spring-boot-example/src/main/java/org/optaplanner/spring/boot/example/domain/TimeTableView.java
@@ -35,4 +35,5 @@ public class TimeTableView {
     public SolverStatus getSolverStatus() {
         return solverStatus;
     }
+
 }
