+++ optaplanner-integration/optaplanner-spring-boot-example/src/main/java/org/optaplanner/spring/boot/example/persistence/RoomRepository.java
@@ -14,16 +14,16 @@
  * limitations under the License.
  */
 
-package org.optaplanner.spring.boot.example.service;
+package org.optaplanner.spring.boot.example.persistence;
 
 import java.util.List;
 
-import org.optaplanner.spring.boot.example.domain.Lesson;
+import org.optaplanner.spring.boot.example.domain.Room;
 import org.springframework.data.repository.PagingAndSortingRepository;
 
-public interface LessonRepository extends PagingAndSortingRepository<Lesson, Long> {
+public interface RoomRepository extends PagingAndSortingRepository<Room, Long> {
 
     @Override
-    List<Lesson> findAll();
+    List<Room> findAll();
 
 }
