+++ optaplanner-integration/optaplanner-spring-boot-example/src/main/java/org/optaplanner/spring/boot/example/persistence/TimeTableRepository.java
@@ -1,0 +1,52 @@
+/*
+ * Copyright 2019 Red Hat, Inc. and/or its affiliates.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package org.optaplanner.spring.boot.example.persistence;
+
+import org.optaplanner.spring.boot.example.domain.Lesson;
+import org.optaplanner.spring.boot.example.domain.TimeTable;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
+
+@Service
+@Transactional
+public class TimeTableRepository {
+
+    @Autowired
+    private TimeslotRepository timeslotRepository;
+    @Autowired
+    private RoomRepository roomRepository;
+    @Autowired
+    private LessonRepository lessonRepository;
+
+    public TimeTable find() {
+        // Occurs in a single transaction, so each initialized lesson references the same timeslot/room instance
+        // that is contained by the timeTable's timeslotList/roomList.
+        return new TimeTable(
+                timeslotRepository.findAll(),
+                roomRepository.findAll(),
+                lessonRepository.findAll());
+    }
+
+    public void save(TimeTable timeTable) {
+        for (Lesson lesson : timeTable.getLessonList()) {
+            // TODO this is awfully naive: optimistic locking will cause issues if called by the SolverManager
+            lessonRepository.save(lesson);
+        }
+    }
+
+}
