+++ optaplanner-integration/optaplanner-spring-boot-example/src/main/java/org/optaplanner/spring/boot/example/poc/impl/solver/DefaultSolverManager.java
@@ -110,7 +110,7 @@ public class DefaultSolverManager<Solution_, ProblemId_> implements SolverManage
         solverJob.reloadProblem(problemSupplier);
     }
 
-    @Override
+//    @Override
     public void addProblemFactChange(ProblemId_ problemId, ProblemFactChange<Solution_> problemFactChange) {
         DefaultSolverJob<Solution_, ProblemId_> solverJob = problemIdToSolverJobMap.get(problemId);
         if (solverJob == null) {
